+++ src/wiki/conf/settings.py
@@ -26,13 +26,28 @@ MARKDOWN_SANITIZE_HTML = getattr(
     'WIKI_MARKDOWN_SANITIZE_HTML',
     True)
 
-#: Arguments for the Markdown instance, for instance a list of extensions to
-#: use.
-#: See: https://pythonhosted.org/Markdown/extensions/index.html
+#: Arguments for the Markdown instance, as a dictionary. The "extensions" key
+#: should be a list of extra extensions to use besides the built-in django-wiki
+#: extensions, and the "extension_configs" should be a dictionary, specifying
+#: the keyword-arguments to pass to each extension.
 #:
-#: To set a custom title for TOC's::
+#: For a list of extensions officially supported by Python-Markdown, see:
+#: https://python-markdown.github.io/extensions/
 #:
-#:    WIKI_MARKDOWN_KWARGS = {'extension_configs': {'toc': _('Contents of this article')}}
+#: To set a custom title for table of contents, specify the following in your
+#: Django project settings::
+#:
+#:     WIKI_MARKDOWN_KWARGS = {
+#:         'extension_configs': {
+#:             'wiki.plugins.macros.mdx.toc': {'title': 'Contents of this article'},
+#:         },
+#:     }
+#:
+#: Besides the extensions enabled by the "extensions" key, the following
+#: built-in django-wiki extensions can be configured with "extension_configs":
+#: "wiki.core.markdown.mdx.codehilite", "wiki.core.markdown.mdx.previewlinks",
+#: "wiki.core.markdown.mdx.responsivetable", "wiki.plugins.macros.mdx.macro",
+#: "wiki.plugins.macros.mdx.toc", "wiki.plugins.macros.mdx.wikilinks".
 MARKDOWN_KWARGS = {
     'extensions': [
         'markdown.extensions.footnotes',
@@ -46,8 +61,8 @@ MARKDOWN_KWARGS = {
         'markdown.extensions.sane_lists',
     ],
     'extension_configs': {
-        'toc': {
-            'title': _('Table of Contents')}},
+        'wiki.plugins.macros.mdx.toc': {'title': _('Contents')},
+    },
 }
 MARKDOWN_KWARGS.update(getattr(django_settings, 'WIKI_MARKDOWN_KWARGS', {}))