+++ src/wiki/core/markdown/mdx/previewlinks.py
@@ -19,3 +19,8 @@ class PreviewLinksTree(Treeprocessor):
                 if not a.get('href').startswith('#'):
                     a.set('target', '_blank')
         return root
+
+
+def makeExtension(*args, **kwargs):
+    """Return an instance of the extension."""
+    return PreviewLinksExtension(*args, **kwargs)