+++ docs/release_notes.rst
@@ -43,6 +43,7 @@ Fixed
  * Use ``simple_tag`` for assignment tag :url-issue:`791` (Raffaele Salmaso)
  * Direct invocation of ``pytest`` fixed (removing ``runtests.py``) :url-issue:`781` (Branko Majic)
  * Line breaks in help texts for macros :url-issue:`851` (Mathias Dannesbo)
+ * Table of contents now has a header by default, and several built-in django-wiki extensions can be configured using ``WIKI_MARKDOWN_KWARGS`` :url-issue:`881` (Mathias Rav)
 
 Deprecated/Removed
 ~~~~~~~~~~~~~~~~~~